# src/main/java/api/bank/dto/UserUpdateRequest.java
package api.bank.dto;

import api.bank.entity.Role;
import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Schema(description = "Запрос на обновление пользователя")
public class UserUpdateRequest {

    @Size(min = 3, max = 50, message = "Имя пользователя должно содержать от 3 до 50 символов")
    @Schema(description = "Имя пользователя")
    private String username;

    @Size(min = 6, message = "Пароль должен содержать минимум 6 символов")
    @Schema(description = "Пароль")
    private String password;

    @Schema(description = "Роль пользователя")
    private Role role;
}